# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products RestApi",
        "version": "0.1"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Returns a list of all products",
                "operationId": "16151fd150adb5c53ccb365b18872b88",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All products"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Create a new product with the provided data",
                "operationId": "39e039020db35ef73fb90039e9a2e0fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/data": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Fetch products with pagination",
                "description": "Fetches products with pagination using DataTables format",
                "operationId": "409d822611154f086ac560143fbd1b16",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Start index for pagination (default: 0)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "0"
                        }
                    },
                    {
                        "name": "length",
                        "in": "query",
                        "description": "Number of items to fetch per page (default: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "draw": {
                                            "description": "Draw counter for DataTables",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "recordsTotal": {
                                            "description": "Total number of records",
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "recordsFiltered": {
                                            "description": "Total number of records after filtering",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "description": "ID unique du produit",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom du produit",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description du produit",
                        "type": "string"
                    },
                    "price": {
                        "description": "Prix du produit",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "Stock du produit",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Date de création du produit",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour du produit",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}